run: set required permissions for OpenDKIM
OpenDKIM by default requires strict permissions on both the directory
of the key files as well as the key files themselves.

If the DKIM key is volume mounted (e.g. not created by `run`), then it
may not have the correct permissions. To fix this, we can always reset
the permissions with chown each time `run` is invoked.

Alternatively, we could use `RequireSafeKeys false` in opendkim.conf
so that the mail is sent regardless of permissions, but this is a less
secure option.