run: improve automatic DKIM private key creation
Run opendkim-genkey with the "-D" ("--directory") option instead of
changing directory manually in a bash subshell.

Also output an informational message when a DKIM private key is being
generated. This is useful to see, especially if you were expecting an
existing file to be found (e.g. if the wrong selector was used).